move Column to internal class of Table and Tree
