+++ src/core/toga/__init__.py
@@ -29,7 +29,6 @@ from .widgets.slider import Slider
 from .widgets.splitcontainer import SplitContainer
 from .widgets.switch import Switch
 from .widgets.table import Table
-from .widgets.tablecolumn import Column
 from .widgets.textinput import TextInput
 from .widgets.timepicker import TimePicker
 from .widgets.tree import Tree
@@ -57,7 +56,6 @@ __all__ = [
     'Box',
     'Button',
     'Canvas',
-    'Column',
     'DetailedList',
     'Divider',
     'Window',
